Add `Document` domain model for julee_example with basic repository. (#18)
The main change here is to add a domain model for `Document` based on my current understanding (mainly from https://github.com/pyx-industries/julee-orchestrator/issues/1 ), together with tests, so we can then mock a repository for the things we do want to start on now.

The main point of interest in that model is that the document itself is a reader (implements `read()`, `seek()` and `tell()` - well, it's really just delegating those to the `content` property which is a content stream) so that the (possibly huge) content does not need to be exposed in the serialisation. Instead, a `content_multihash` is included in the serialisation and the actual content can be streamed when needed. I'm not sure yet if this is valid, but it seems like the only sane way to handle documents which could be large.

I also added a first `DocumentRepository` protocol with just get, update and store (and generate_document_id, similar to the sample, so that the store call is idempotent).


Sadly, a decent amount of time in this PR was spent pruning all the extra cruft being generated - can we update the fun-police to call out small, focused changes or something (and while we're there, I'd like to add something about the "You're absolutely right", "That's a really brilliant insight" - or I'm going to develop narcissism... worry what it's doing to society already, but it creates happier paying customers I'm sure).