Add MinioDocumentRepository implementation (#19)
This PR follows on from #18 and adds an implementation of the `MinioDocumentRepository` mainly so I can then ensure the design works fine. It's actually only half of the change - I've split the branch into the implementation and the tests and will push the second half which contains the tests with some small updates.

The main architectural point here is that the actual data from a document is stored via a content-addressable key (the multihash) and *always* uses a streaming type (ie. no reads into memory by default).