Add Assembly domain model with tests and example. (#22)
Also adds the KnowledgeServiceQuery / Extractor domain model.

The key points for this change is that the Assembly has a jsonschema defining the data required to be assembled, and queries which can be applied to different parts of the data, as required. Publishing the assembly would then involve rendering it as needed (with different options possible).