Add `Assembly` and `AssemblyIteration` models. (#25)
This PR adds the `Assembly` and `AssemblyIteration` models, noting in the docstrings that the iteration model will later include scorecards for the output document.

Note: there are other short-cuts here, such as the assembly's input being a document ID (from the storage, whereas I assume later it'll be from a knowledge service), but please check that the basic modeling is inline with what you imagine.