Implement anthropic knowledge service (#34)
This PR implements both `register_file` and `execute_query` for the Anthropic implementation of knowledge service.

While there, I updated the knowledge service query type to include query metadata that may be specific to the knowledge service and used that in execute query.

I have a local test script that ensures this actually works too, but haven't included that here (as I want to include a better demo script once we have the assembly executing from start to finish).

The next PR will tweak the anthropic implementation to return a file rather than a string in the conversation, I think (matching our query expectations for operating on a document to produce a document).