Proper query execution for anthropic knowledge service (#35)
This PR follows #34 and updates the query execution in the anthropic knowledge service implementation.

I wasn't able to get the anthropic API to return a file (apparently we can do that with claude code, but not the normal messages API), but since you can constrain the response by adding assistant messages to your query (so that the response just continues from there), we can get a similar effect (ie. the response is *just* the content of the expected file).

So, assuming that the assistant prompt is a normal feature across LLMs, I've added it to the KnowledgeServiceQuery domain model (but if it's not a normal feature, we can move it into the `query_metadata` instead).

To demonstrate this working, I've included the generated example I've been playing with, which is for a single query only. It registers two files with the knowledge service (the meeting transcript, the json schema for the extracted meeting minutes data) and a user prompt followed by an assistant prompt which constrains the result. Note the output below doesn't display the assistant prompt that I've used, but you can see it in the code, it's just "Looking at the meeting transcript, here's the formatted JSON summary that satisfies the provided schema, without surrounding ```json ... ``` markers:"