Split out transcript and schema from demo. Add args. (#41)
Follows #40, just extracts the example data from the example script (and adds the option of passing them via the CLI).

The reason for this is just as a prequel to add a couple of render templates, since for a quick demo, a rendered output or two is going to be more interesting? (though it's not really doing much - just rendering jinja2, wdyt?)