Add second meeting minutes assembly spec example. (#42)
Same minutes, same schema, but different voice. I realise this isn't a great extraction example, as it's more of a transformation, but useful to see how the queries can be modified and to demonstrate different specifications with the same data.

This PR adds a new `meeting_minutes_assembly_spec-sk8r.json` specification, which has the same schema and can be run with the same meeting minutes, but uses different queries that require skateboard community language where appropriate. As above, normally an extraction would avoid this, but since we don't yet have transformations, I just thought it interesting to see how the extraction queries can be used. Pretty over-the-top, but funny ("should stick the landing by Friday", ""Mike estimated 2 weeks for frontend grinding")... and there's a lot of "tricks" in there...