Add `Policy` domain model (#43)
This PR adds a `Policy` domain model.

I've tried to capture the idea of "transformation" vs "non-transformation" (validation-only) dependent on the optional `transformation_queries` attribute. If it's empty, then it's a validation-only policy (with the `validation_scores` field always required), whereas if `transformation_queries` is populated, then we know its a transformation policy (and can run the validation scores before and after the transformation, for example).