Add policy repository implementations (minio, memory). (#44)
Follows #43.

Adds a policy repository protocol together with minio and memory implementations.

We have quite a few repository protocols that implement a similar interface, just with different return types. I'll take a quick look if we can use python generics for this, to remove code. Similar for the minio and memory implementations (which all do the same thing, with different types). I'm guessing down the line we'll need differences between the repositories for different objects, but until we do...