Add a base repository protocol to DRY up repository protocols. (#45)
Follows #44. As mentioned there, I noticed that we're basically doing the same protocol for all repositories (get, save, generate_id), just returning different types.

This PR adds a base protocol using python generics, so that each specific repository protocol is just a concrete protocol of that generic base.

It required removing the two specific methods, which didn't need to be specific (eg. `DocumentRepository` had an `update` and `store`, which were doing almost the same thing, which could be handled by the save implementation easily).

I'll next do something similar for the minio and memory implementations, which are also duplicating a lot of code.