Add support for validating transformation policies (#50)
Follows #49 and updates the `validate_document` function to support transformation policies as well.

If the initial validation passes, or if it fails and there are no transformations, it returns the result, otherwise it runs the transform(s) for this policy (we only envisage 1 per policy, but currently it's a list as per previous discussion) before it runs the validation again, records and returns the result.

The next PR will attempt to update the example demo to use a transformation policy.