Add activities and workflow proxies for existing repositories (#52)
This PR:
- Renames the `temporal_repository` decorator to `temporal_activity_registration` (since that is what is doing - not really specific to repositories). I've still exported it with the old name so as to not pollute this diff with fixing existing call-sites - I'll do that separately.
- Adds a new `temporal_workflow_proxy` decorator which automatically creates the workflow proxy version of a class.

So the first one takes the protocol methods on a concrete implementation (eg. minio) and decorates them as temporal activities. The second one creates a class that ensures that when the same protocol method is called, it delegates to calling the temporal workflow activity (registered by the first decorator above).

I got to this point because, while getting my head around what the proxy was for and implementing them for each repository, I found I was repeating the same code for each repository, so again looked for a way to reduce boilerplate. If it works (I've not yet tested it), it'll make it much clearer and simpler to maintain existing temporal repos as well as add new ones, reducing our 2 layers to 1 (previously reduced from 3 to 2 by the first decorator).

If you're OK with this approach, the next PR will:
- fix the TODO I left about existing call-sites to `temporal_repository` decorator, and
- include unit tests for the new decorator.
Then the one after that will actually use the new proxy classes in a temporal workflow and test that (using temporal's test helper for workflows).