Reorganise temporal activities and proxies in sample to match julee example. (#55)
Following on from #54, and in particular, [this comment](https://github.com/monkeypants/plays_with_temporal/pull/54/files#r2366596534).

This PR just updates the way the temporal activities and proxies are organised, to match the way I've done them in the julee_example, and updates the docs to document that.

If you disagree, happy to switch the julee example to match sample and revert the doc change.

The reason for this is:

- temporal activities were all defined together *outside* of the temporal directory (not sure of the reason for that part), but more importantly,
- the activities as well as the function proxies which call those activities are pretty intimately related - the two decorators share the `_discover_protocol_methods` function to ensure that the activities automatically created by the one decorator match the activities called by the functions of the other.
- Since the decorators make the class definitions empty, having them both in the same file, organised within `temporal` with a file name matching the repository name (like all the other directories) seems to be much clearer, to me.

See what you think.