Fix decorator to use correct domain types. (#56)
I was finding that temporal workflows were failing because they were being passed python dicts, rather than Pydantic domain models, which caused `AttributeError`'s when trying to access Pydantic model attributes using dot notation.

It took a while to debug, but it came down to the  `@temporal_workflow_proxy` decorator in `util/repos/temporal/decorators.py` being applied to the generic `class BaseRepository(Protocol[T]):`  model that I created since all repositories had the same methods.

This left the decorator unable to know the actual return type (it just knew it as the generic `T`), and so when temporal's pydantic plugin tried to determine the type to convert the json to a pydatic model, it couldn't and so was just returning the json.

I initially fixed this by updating the decorator signature to include the domain type, but then realised that we already have that information in the inheritance chain (eg. a `WorkflowAssemblySpecificationRepository` inherits from `BaseRepository(Protocol[AssemblySpecification])`, so this PR updates the decorator to pull out the `AssemblySpecification` and update the return types for the decorated methods.