Enable saving a Document with content_string for small payloads. (#57)
This is another prequel branch for the temporal integration that I'm working on.

## The problem

It is a use-case (extract-assemble-data.py) which calls out to the different knowledge services to extract different sections of the required data, and then the use-case is the one stitching those together to form the final document. Since the use-case deals only with temporal proxies (ie. no concrete implementations), it had no way of *saving* the resulting document content (since the content stream is not included in the serialisation/deserialisation which is required when calling activities like save from a workflow).

## The solution
I tried a couple of different solutions to this, but this one seems the cleanest: include a `content_string` optional attribute on the document model, for use only when saving documents with small content (a few kb) from use-cases/workflows. Any document saves from an activity will be using concrete implementations and so can save with the normal content stream (as there's no temporal serialisation happening on the save, in that case).