Knowledge service calls always take config. (#58)
So: I have the temporal examples working locally now, but due to the investigative nature of the process, it's nearly 4k of diff.

I've looked for ways to break it up such that tests pass with each, but it's tricky, so I'm instead going to propose a series of 200-500 line PRs which have a concept to review, but each won't pass all checks. I'll be merging them down into the workflow-4.6.0b which this PR targets, rather than main, so we can ensure once everything is reviewed and in that one branch, it's all passing before landing in the main branch.

This PR's main change is to update the individual knowledge service protocol methods: `register_file` and `execute_query` to accept the knowledge service config (rather than it being passed as an init arg). This is so we can know at run-time which knowledge service implementation should be used for the call. Previously I'd planned to instantiate each knowledge service with the config, but the activities we're defining on classes are effectively class methods / stand-alone functions given the way an activity needs to be retried, so storing instance state makes no sense (I actually wonder whether using classes rather than just modules with functions, is actually benefiting us - but not worth devoting time to now).

There are a few other small changes, for which I'll leave comments inline.