Add the temporal activities and proxies. (#59)
Following on from #58 , this PR:
- Updates the memory implementation to match the updated protocol of #58, and
- adds the temporal activities and proxy (with shared base activity names).

While reading through this diff, I've seen an issue (not functional, but architectural) where I should be using dependency injection rather than instantiating concrete classes in the __init__ of the TemporalKnowledgeService. I've left a `TODO(temporal)` there, as I can't fix it confidently until all the code in this series is reviewed and in the one branch.

If it's OK, I'll do that with any other feedback too (leave a TODO do a separate PR fixing those once the working code is reviewed and in the one branch).

The TemporalKnowledgeService just wraps the non-temporal one, but retrieves documents to pass through, since with temporal serialisation before an activity is running, we lose the content stream.