Allow temporal serialization of Document to not have content. (#64)
Follows #63 , which added the new example worker and client script, but it failed to run due to the new domain validation we added on Document, that either the `content` (stream) or `content_string` is set. While that makes perfect sense when *creating* a Document, it does not when temporal is serialising a document. For example, if we create the document with a `content` (a stream), we save it - (stream gets saved to the `documet-content` bucket using content-addressable hash), then temporal serialises it to send over the wire... when it does so, it should *not* have either set.

So the fix here is using our own modified data converter, which just adds a bit of context which the domain model can check for, so it allows the empty content when deserialising from temporal.