Workflow reviewed changes (#60)
This branch is a collection of reviewed PRs, each of which wasn't passing on its own as I had to divide the investigative work up into reviewable chunks.

You can find the details in PRs #58, #59, #61, #62, #63 and #64.

This PR brings those all together (with tests passing), adding an example which can be run demonstrating the same extraction and assembly of meeting minutes (current example) but handled by temporal activities.

Run the temporal worker in one terminal with:

``` PYTHONLOGLEVEL=debug python -m julee_example.worker 2>&1 | tee debug.txt ```

and in another terminal, the example CLI client which sets the initial data (a stub "capture step") and requests the extract-and-assemble workflow, with:
```python -m julee_example.examples.client_example```