Add profiles to compose. Debug and work-around get_many (#69)
Initially this was just adding profiles to the docker compose together with a dockerfile for the worker, so that I could test the julee worker more easily. But the julee worker was erroring.

After too much debugging, it came down to the switch to `get_many` that I'd recently landed - it causes serialization issues with temporal. I didn't want to put more time into investigating it now, so without an issue tracker here, I've left a decent TODO with the details.

This PR also includes a new decorator which helps identify serialization issues with clearer log messages for better debugging (it checks that the input types for activities called from temporal are the expected domain models, not dicts).

So to start the julee example, it's just:

```console
$ docker compose --profile julee up
```

Or to reset (deleting volumes so no stale tasks - helpful when debugging):

```console
$ docker compose --profile julee down -v && docker compose --profile julee up -d
```

With it running, we can trigger an example workflow with:

```console
python -m julee_example.examples.client_example
```

I'm going to follow-up with another PR which deletes the old examples (now that it works with temporal), possibly improve the example for a demo, and eventually add a CI e2e test to ensure it doesn't get broken by another commit (like the switch to get_many did).