Update temporal client example to use policy validation flow as well (#70)
The client example now calls the policy validation on the result of the extract-assemble workflow.

I had to update the policy validation side of things to work with temporal (similar changes to what was needed for the extract-assemble workflow - creating proxies, registering activity names - all trivial and without boilerplate due to the decorators :) ).

Also:
- Deleted a bunch of old CLI example code and data
- Updated the populate script to populate the policy data to minio as well,
- updated the client script to (manually) call the policy validation on the result (it's not a workflow decision yet)
- Updated the anthropic knowledge service to check if the content-type of a file being registered is json, and if so, register it as plain text (since anthropic's file services is limited in that way - plain text or pdf only).
- generated the actual workflow for policy validation