Add request object for creating an AssemblySpecification. (#73)
Follows #72 and adds an initial FastAPI pydantic request model for creating an AssemblySpecification. I think the AssemblySpecification and KnowledgeServiceQuery are (currently) the only models we'll need to access via the API for a demo, since we can pre-populate knowledge service configurations and policies?

The main difference between what I've done here and what is in sample is just trying to ensure the `CreateAssemblySpecificationRequest` is DRY, so it re-uses the descriptions from the domain model (though we can update them to be API-specific in needed) and delegates all validation to the existing domain model's validation class methods.