Add api endpont for listing knowledge service queries (#74)
This PR adds an API model for creating a KnowledgeServiceRequest and adds the first API endpoint, for listing knowledge service requests.

It follows the pattern in `sample` for the dependency injection. But one difference is that I've used the fastapi-pagination package to automatically add pagination to results.

Later down the track, we may also want to use something like https://github.com/zmievsa/cadwyn for versioning, but too early yet (IMO).