Add demo UI (#79)
Follows #78 , adds a julee_example/ui directory with the boilerplate for a Vite+React+shadcn frontend app for demoing Julee assemblies.

They say a picture is worth a thousand words, how about 11.5 thousand lines?

It doesn't do anything at the moment, other than display the above (which uses the api only to check the counts of queries and assemblies in the system). This PR is just an agent using the package.json of the existing stuf project to build an SPA with a similar setup, then me tweaking a few things in the setup (like central config). The actual functionality, enabling us to define new assemblies and run them, inspect the evidence of a completed assembly etc will follow piecemeal.