Add demo UI 4 (#82)
Follows #81, this PR uses a fastAPI lifespan to run a special use case on startup, populating the initial knowledge services.

With this PR, we can now select from the available knowledge services when creating a query.

As mentioned previously, I'd like to move the definition of queries into a more integrated flow when creating a specification soon.