React downgrade and small UX fixes (#83)
Follows #82.

I spent some time investigating options for specifying the data (json schema) visually, and found the excellent https://github.com/cern-sis/react-formule , which I plan to use, but it's not compatible with React 19 currently. The only reason I'd upgraded to 19 was when there was an issue with `forwardRef`'s not being defined in the shadcn components and I saw it was because `forwardRef`s have been deprecated in react 19, so upgrading seemed the best way forward (even though the Stuf project is still on 18). But to use Formule, I need to stay on 18 anyway, so this PR downgrades again, and adds the missing `forwardRef`s to the shadcn-generated components.

Next up, I'll create a basic form for creating assembly specifications, that can then be improved with formule etc.

This PR also fixes a small UX issue where the example prompts were overrunning their card boundaries.