Integrate react-formule for jsonschema definitions (#86)
This PR provides initial integration with react-formule, to enable easier definition of the assembly data to be extracted, though the next couple of PRs will continue to make this simpler.

The package lock is 3000 lines, and there are less relevant changes in the src/components/ui (the shadcn components). The main focus is on the JsonSchemaEditor.tsx and AssemblySpecificationForm.tsx .

The initial UX segment looks like the following:
<img width="945" height="855" alt="image" src="https://github.com/user-attachments/assets/36f1d8f8-6d85-40d9-893f-93efd55c2985" />

But the next couple of branches following will:
- Customise the available fields to those we're limited to for json schema (object, list, text and number)
- allow the user to use prompts to create the data format in the UX