Adds a custom ai chat to help generate the data format with diff (#87)
Follows #86. This PR adds a custom ai chat (the integrated one was too focused on creating forms rather than data) which makes creating the data format simple. Here are some screenshots demoing how it works:

## Asking for a data format

<img width="989" height="844" alt="Screenshot from 2025-10-28 11-09-40" src="https://github.com/user-attachments/assets/84e5af22-1cc6-43bf-ac6a-967960e9552a" />

## Reviewing the result

<img width="989" height="844" alt="Screenshot from 2025-10-28 11-10-06" src="https://github.com/user-attachments/assets/3521bd25-7d44-4d5d-98b6-bcb5b34956e6" />

## Asking for a change to the data format

<img width="989" height="844" alt="Screenshot from 2025-10-28 11-10-35" src="https://github.com/user-attachments/assets/80116685-6c3b-430c-bc96-c493c47144fb" />

## Reviewing the result

<img width="989" height="844" alt="Screenshot from 2025-10-28 11-10-55" src="https://github.com/user-attachments/assets/b1892439-bc1e-4f5c-bd96-1b4a992bfbcd" />


Note: it's currently using the API key in the frontend code, which is not the way it will remain beyond a demo... I'll update to proxy that call via the backend when we've time.

The next PR will tidy up the field selections with a custom component.