Add custom select or edit so we only see fields that we need (#88)
Follows #87 - this PR adds a custom field selector which just has the fields we're interested in (object, list, text, number), while still allowing editing of those.

<img width="957" height="1022" alt="Screenshot from 2025-10-28 14-06-59" src="https://github.com/user-attachments/assets/e8e081f7-5bf3-4232-ba94-9fe1ab85d2a8" />

There's a bit of a unclean hack catching an error due to being unable to manage the default formule state, but it works for a demo (we should create an upstream issue when time).