Add KnowledgeServiceQueryDisplay component. (#90)
This PR follows #89 and replaces the text KnowledgeServiceQuery field with something which will be simpler to interact with. At the moment it just displays as:

<img width="949" height="543" alt="Screenshot from 2025-10-28 17-20-15" src="https://github.com/user-attachments/assets/4d05154d-4351-46de-a69e-362acfa79032" />

but when queries are associated with parts of the data, they will display here neatly, mapping the field name to the query name, and allowing clicking on one to switch the field editor (above on the form) to that field, where the association can be edited or modified.

The next PR will do the latter, enabling associating a query with part of the data.