Enable associating query with data in schema (#91)
Follows #90, this PR adds the ability to associate a query with a data element in the schema.

Not all navigation works - I'm going to pause this effort to now focus on running an assembly next, now that we can at least demo creating one in the UX (I might add a fixture for one too).

<img width="925" height="974" alt="Screenshot from 2025-10-29 13-48-45" src="https://github.com/user-attachments/assets/f247a06e-deb7-46b8-aeac-21dc6e257a8d" />