Add workflow_id to `Assembly` model (#98)
Follows #97. Ref: https://github.com/gs-gs/pyx-rbtp/issues/85

This PR adds the `workflow_id` to the `Assembly` model. The original intent was that I was wanting to be able to redirect to the assembly when an assembly specification is run, and stream data from the workflow to provide updates. BUT I think there's an architectural change we'll need to make post-demo, so I'm not going to push forward with this for now and will instead just tidy up the current work.

The issue is that running an assembly triggers a (temporal) workflow, which runs asynchronously, and it is within this workflow execution that the `Assembly` is created - so we cannot redirect to an `Assembly` currently. I'm not yet sure if we want to
1. generate the pending assembly with the API call so that we can redirect to it and watch for changes (means generating the assembly ID in the API call rather than the workflow), or
2. abstract the workflow execution to a model (ie. not depend on temporal) and redirect to that for the execution.

At least with this PR, we are now able to find the assembly once the workflow completes (not yet sure if I'll use that for now).